Move MinHeap.java and MaxHeap.java to the "FixedArrayImplementation" folder.
- This is because these min heaps and max heaps are implemented using fixed arrays.
