+++ src/DataStructures/CustomDataStructures/MinHeap/FixedArrayImplementation/MinHeap.java
@@ -1,8 +1,8 @@
-package DataStructures.CustomDataStructures.MaxHeap;
+package DataStructures.CustomDataStructures.MinHeap.FixedArrayImplementation;
 
 import java.util.Arrays;
 
-public class MaxHeap {
+public class MinHeap {
 
     // source: https://www.youtube.com/watch?v=t0Cq6tVNRBA min heaps
 
@@ -11,9 +11,9 @@ public class MaxHeap {
     private int[] items;
 
     /**
-     * Create a new max heap.
+     * Create a new min heap.
      */
-    public MaxHeap() {
+    public MinHeap() {
         size = 0;
         capacity = 10;
         items = new int[capacity];
@@ -21,14 +21,14 @@ public class MaxHeap {
     }
 
     /**
-     * Return true iff this max heap contains no items.
+     * Return true iff this min heap contains no items.
      */
     public boolean isEmpty() {
         return size == 0;
     }
 
     /**
-     * Add the item to this max heap.
+     * Add the item to this min heap.
      */
     public void add(int item) {
         ensureExtraCapacity();
@@ -38,21 +38,21 @@ public class MaxHeap {
     }
 
     /**
-     * Return the top of this max heap.
+     * Return the top of this min heap.
      */
     public int peek() {
         if (size == 0) {
-            throw new IllegalStateException("There are no items in this max heap.");
+            throw new IllegalStateException("There are no items in this min heap.");
         }
         return items[0];
     }
 
     /**
-     * Remove and return the top of this max heap.
+     * Remove and return the top of this min heap.
      */
     public int poll() {
         if (size == 0) {
-            throw new IllegalStateException("You cannot delete from an empty max heap.");
+            throw new IllegalStateException("You cannot delete from an empty min heap.");
         }
 
         int item = items[0];
@@ -64,21 +64,21 @@ public class MaxHeap {
     }
 
     /**
-     * Return a string representation of this max heap.
+     * Return a string representation of this min heap.
      */
     public String toString() {
-        String maxHeapString = "";
+        String minHeapString = "";
         for (int i = 0; i < size; i++) {
-            maxHeapString += items[i] + " ";
+            minHeapString += items[i] + " ";
         }
 
-        maxHeapString = maxHeapString.trim();
+        minHeapString = minHeapString.trim();
         if (size >= 1) {
-            maxHeapString += " ";
+            minHeapString += " ";
         }
 
-        maxHeapString += "Size: " + size;
-        return maxHeapString;
+        minHeapString += "Size: " + size;
+        return minHeapString;
     }
 
     private int getLeftChildIndex(int parentIndex) {
@@ -114,35 +114,35 @@ public class MaxHeap {
     private void ensureExtraCapacity() {
         if (size == capacity) {
             // copy over items into a bigger array
-            items = Arrays.copyOf(items, capacity * 2);
+            items = Arrays.copyOf(items, capacity * 2); // copy over items into a bigger array
             capacity *= 2;
         }
     }
 
-    private int getBiggerChildIndex(int index) {
-        int biggerChildIndex = getLeftChildIndex(index);
-        if (hasRightChild(index) && rightChild(index) > leftChild(index)) {
-            biggerChildIndex = getRightChildIndex(index);
+    private int getSmallerChildIndex(int index) {
+        int smallerChildIndex = getLeftChildIndex(index);
+        if (hasRightChild(index) && rightChild(index) < leftChild(index)) {
+            smallerChildIndex = getRightChildIndex(index);
         }
-        return biggerChildIndex;
+        return smallerChildIndex;
     }
 
     private void heapifyDown() {
         int index = 0; // start with root
         while (hasLeftChild(index)) {  // no left child => no right child
-            int biggerChildIndex = getBiggerChildIndex(index);
-            if (items[index] > items[biggerChildIndex]) {
+            int smallerChildIndex = getSmallerChildIndex(index);
+            if (items[index] < items[smallerChildIndex]) {
                 break; // heap elements now in order
             } else {
-                swap(index, biggerChildIndex);
+                swap(index, smallerChildIndex);
             }
-            index = biggerChildIndex;
+            index = smallerChildIndex;
         }
     }
 
     private void heapifyUp() {
         int index = size - 1; // start with last element
-        while (hasParent(index) && parent(index) < items[index]) {
+        while (hasParent(index) && parent(index) > items[index]) {
             swap(getParentIndex(index), index);
             index = getParentIndex(index);
         }
